refactor(vo/service): 将计数类型从Integer改为Long
修改CommunityNoteVO中的favoriteCount和commentCount字段类型，以及相关服务接口和实现类中的计数返回类型，从Integer改为Long以支持更大的数值范围

+++ backend/src/main/java/com/wyl/note/service/NoteCommentService.java
@@ -16,5 +16,5 @@ public interface NoteCommentService extends IService<NoteComment> {
 
     List<NoteCommentVO> getCommentsByNoteId(Long noteId);
 
-    Integer getCommentCount(Long noteId);
+    Long getCommentCount(Long noteId);
 }
